services/identity-service/requirements.txt
fastapi
uvicorn
motor
python-jose
passlib[bcrypt]
python-dotenv
email-validator
pandas 
openpyxl 
python-multipart
argon2-cffi
xlrd
